--- src/main/java/kitchenpos/application/TableGroupService.java
@@ -1,35 +1,0 @@
-package kitchenpos.application;
-
-import kitchenpos.application.validator.TableGroupValidator;
-import kitchenpos.domain.TableGroup;
-import kitchenpos.dto.TableGroupResponse;
-import kitchenpos.dto.TableGroupSaveRequest;
-import kitchenpos.repository.TableGroupRepository;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@Transactional(readOnly = true)
-public class TableGroupService {
-
-    private final TableGroupValidator tableGroupValidator;
-    private final TableGroupRepository tableGroupRepository;
-
-    public TableGroupService(final TableGroupValidator tableGroupValidator,
-                             final TableGroupRepository tableGroupRepository) {
-        this.tableGroupValidator = tableGroupValidator;
-        this.tableGroupRepository = tableGroupRepository;
-    }
-
-    @Transactional
-    public TableGroupResponse create(final TableGroupSaveRequest request) {
-        TableGroup savedTableGroup = tableGroupRepository.save(new TableGroup(request.getOrderTables()));
-        return new TableGroupResponse(savedTableGroup);
-    }
-
-    @Transactional
-    public void ungroup(final Long tableGroupId) {
-        TableGroup savedTableGroup = tableGroupRepository.getById(tableGroupId);
-        savedTableGroup.ungroup(tableGroupValidator.validate(savedTableGroup));
-    }
-}
