--- src/main/java/kitchenpos/application/validator/TableValidator.java
@@ -1,9 +1,0 @@
-package kitchenpos.application.validator;
-
-import java.util.function.Supplier;
-
-@FunctionalInterface
-public interface TableValidator {
-
-    Supplier<Boolean> validate(final Long orderTableId);
-}
