--- src/main/java/kitchenpos/dao/JpaOrderTableDao.java
@@ -1,7 +1,0 @@
-package kitchenpos.dao;
-
-import kitchenpos.domain.OrderTable;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JpaOrderTableDao extends JpaRepository<OrderTable, Long>, OrderTableDao {
-}
