--- src/main/java/kitchenpos/dao/JpaTableGroupDao.java
@@ -1,7 +1,0 @@
-package kitchenpos.dao;
-
-import kitchenpos.domain.TableGroup;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JpaTableGroupDao extends JpaRepository<TableGroup, Long>, TableGroupDao {
-}
