--- src/main/java/kitchenpos/domain/OrderTable.java
@@ -1,85 +1,0 @@
-package kitchenpos.domain;
-
-import java.util.Objects;
-import java.util.function.Supplier;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-public class OrderTable {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "table_group_id")
-    private Long tableGroupId;
-
-    @Column(name = "number_of_guests")
-    private int numberOfGuests;
-
-    @Column(name = "empty")
-    private boolean empty;
-
-    protected OrderTable() {
-    }
-
-    public OrderTable(final Long tableGroupId, final int numberOfGuests, final boolean empty) {
-        this.tableGroupId = tableGroupId;
-        this.numberOfGuests = numberOfGuests;
-        this.empty = empty;
-    }
-
-    public void group(final Long tableGroupId) {
-        this.tableGroupId = tableGroupId;
-        this.empty = false;
-    }
-
-    public void ungroup() {
-        this.tableGroupId = null;
-        this.empty = false;
-    }
-
-    public void changeNumberOfGuests(final int numberOfGuests) {
-        if (numberOfGuests < 0) {
-            throw new IllegalArgumentException();
-        }
-
-        if (isEmpty()) {
-            throw new IllegalArgumentException();
-        }
-
-        this.numberOfGuests = numberOfGuests;
-    }
-
-    public void changeEmpty(final Supplier<Boolean> condition, final boolean empty) {
-        if (Objects.nonNull(tableGroupId)) {
-            throw new IllegalArgumentException();
-        }
-
-        if (condition.get()) {
-            throw new IllegalArgumentException();
-        }
-
-        this.empty = empty;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public Long getTableGroupId() {
-        return tableGroupId;
-    }
-
-    public int getNumberOfGuests() {
-        return numberOfGuests;
-    }
-
-    public boolean isEmpty() {
-        return empty;
-    }
-}
