--- src/main/java/kitchenpos/dto/OrderTableChangeEmptyRequest.java
@@ -1,17 +1,0 @@
-package kitchenpos.dto;
-
-public class OrderTableChangeEmptyRequest {
-
-    private boolean empty;
-
-    private OrderTableChangeEmptyRequest() {
-    }
-
-    public OrderTableChangeEmptyRequest(final boolean empty) {
-        this.empty = empty;
-    }
-
-    public boolean isEmpty() {
-        return empty;
-    }
-}
