--- src/main/java/kitchenpos/dto/OrderTableChangeNumberOfGuestsRequest.java
@@ -1,17 +1,0 @@
-package kitchenpos.dto;
-
-public class OrderTableChangeNumberOfGuestsRequest {
-
-    private int numberOfGuests;
-
-    private OrderTableChangeNumberOfGuestsRequest() {
-    }
-
-    public OrderTableChangeNumberOfGuestsRequest(final int numberOfGuests) {
-        this.numberOfGuests = numberOfGuests;
-    }
-
-    public int getNumberOfGuests() {
-        return numberOfGuests;
-    }
-}
