+++ src/main/java/kitchenpos/table/application/TableService.java
@@ -1,0 +1,56 @@
+package kitchenpos.table.application;
+
+import static java.util.stream.Collectors.*;
+
+import kitchenpos.table.validator.TableValidator;
+import kitchenpos.table.domain.OrderTableDao;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.dto.OrderTableChangeEmptyRequest;
+import kitchenpos.table.dto.OrderTableSaveRequest;
+import kitchenpos.table.dto.OrderTableResponse;
+import kitchenpos.table.dto.OrderTableChangeNumberOfGuestsRequest;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Service
+@Transactional(readOnly = true)
+public class TableService {
+
+    private final OrderTableDao orderTableDao;
+    private final TableValidator tableValidator;
+
+    public TableService(final OrderTableDao orderTableDao, final TableValidator tableValidator) {
+        this.orderTableDao = orderTableDao;
+        this.tableValidator = tableValidator;
+    }
+
+    @Transactional
+    public OrderTableResponse create(final OrderTableSaveRequest request) {
+        OrderTable orderTable = orderTableDao.save(request.toEntity());
+        return new OrderTableResponse(orderTable);
+    }
+
+    public List<OrderTableResponse> list() {
+        return orderTableDao.findAll()
+                .stream()
+                .map(OrderTableResponse::new)
+                .collect(toList());
+    }
+
+    @Transactional
+    public OrderTableResponse changeEmpty(final Long orderTableId, final OrderTableChangeEmptyRequest request) {
+        final OrderTable savedOrderTable = orderTableDao.getById(orderTableId);
+        savedOrderTable.changeEmpty(tableValidator.validate(savedOrderTable.getId()), request.isEmpty());
+        return new OrderTableResponse(savedOrderTable);
+    }
+
+    @Transactional
+    public OrderTableResponse changeNumberOfGuests(final Long orderTableId,
+                                                   final OrderTableChangeNumberOfGuestsRequest request) {
+        OrderTable savedOrderTable = orderTableDao.getById(orderTableId);
+        savedOrderTable.changeNumberOfGuests(request.getNumberOfGuests());
+        return new OrderTableResponse(savedOrderTable);
+    }
+}
