+++ src/main/java/kitchenpos/table/domain/JpaOrderTableDao.java
@@ -1,0 +1,6 @@
+package kitchenpos.table.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JpaOrderTableDao extends JpaRepository<OrderTable, Long>, OrderTableDao {
+}
