+++ src/main/java/kitchenpos/table/domain/JpaTableGroupDao.java
@@ -1,0 +1,6 @@
+package kitchenpos.table.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JpaTableGroupDao extends JpaRepository<TableGroup, Long>, TableGroupDao {
+}
