+++ src/main/java/kitchenpos/table/domain/TableGroupDao.java
@@ -1,0 +1,15 @@
+package kitchenpos.table.domain;
+
+import kitchenpos.table.domain.TableGroup;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface TableGroupDao {
+
+    TableGroup save(final TableGroup entity);
+
+    Optional<TableGroup> findById(final Long id);
+
+    List<TableGroup> findAll();
+}
