+++ src/main/java/kitchenpos/table/dto/OrderTableChangeEmptyRequest.java
@@ -1,0 +1,17 @@
+package kitchenpos.table.dto;
+
+public class OrderTableChangeEmptyRequest {
+
+    private boolean empty;
+
+    private OrderTableChangeEmptyRequest() {
+    }
+
+    public OrderTableChangeEmptyRequest(final boolean empty) {
+        this.empty = empty;
+    }
+
+    public boolean isEmpty() {
+        return empty;
+    }
+}
