+++ src/main/java/kitchenpos/table/dto/OrderTableChangeNumberOfGuestsRequest.java
@@ -1,0 +1,17 @@
+package kitchenpos.table.dto;
+
+public class OrderTableChangeNumberOfGuestsRequest {
+
+    private int numberOfGuests;
+
+    private OrderTableChangeNumberOfGuestsRequest() {
+    }
+
+    public OrderTableChangeNumberOfGuestsRequest(final int numberOfGuests) {
+        this.numberOfGuests = numberOfGuests;
+    }
+
+    public int getNumberOfGuests() {
+        return numberOfGuests;
+    }
+}
