+++ src/main/java/kitchenpos/table/dto/TableGroupResponse.java
@@ -1,0 +1,38 @@
+package kitchenpos.table.dto;
+
+import static java.util.stream.Collectors.*;
+
+import java.time.LocalDateTime;
+import java.util.List;
+import kitchenpos.table.domain.TableGroup;
+
+public class TableGroupResponse {
+
+    private Long id;
+    private LocalDateTime createdDate;
+    private List<OrderTableResponse> orderTables;
+
+    private TableGroupResponse() {
+    }
+
+    public TableGroupResponse(final TableGroup tableGroup) {
+        this.id = tableGroup.getId();
+        this.createdDate = tableGroup.getCreatedDate();
+        this.orderTables = tableGroup.getOrderTables()
+                .stream()
+                .map(OrderTableResponse::new)
+                .collect(toList());
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public LocalDateTime getCreatedDate() {
+        return createdDate;
+    }
+
+    public List<OrderTableResponse> getOrderTables() {
+        return orderTables;
+    }
+}
