+++ src/main/java/kitchenpos/table/ui/TableGroupRestController.java
@@ -1,0 +1,33 @@
+package kitchenpos.table.ui;
+
+import kitchenpos.table.application.TableGroupService;
+import kitchenpos.table.dto.TableGroupResponse;
+import kitchenpos.table.dto.TableGroupSaveRequest;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+
+@RestController
+@RequestMapping("/api/table-groups")
+public class TableGroupRestController {
+
+    private final TableGroupService tableGroupService;
+
+    public TableGroupRestController(final TableGroupService tableGroupService) {
+        this.tableGroupService = tableGroupService;
+    }
+
+    @PostMapping
+    public ResponseEntity<TableGroupResponse> create(@RequestBody final TableGroupSaveRequest request) {
+        final TableGroupResponse created = tableGroupService.create(request);
+        final URI uri = URI.create("/api/table-groups/" + created.getId());
+        return ResponseEntity.created(uri).body(created);
+    }
+
+    @DeleteMapping("/{tableGroupId}")
+    public ResponseEntity<Void> ungroup(@PathVariable final Long tableGroupId) {
+        tableGroupService.ungroup(tableGroupId);
+        return ResponseEntity.noContent().build();
+    }
+}
