+++ src/main/java/kitchenpos/table/validator/TableValidator.java
@@ -1,0 +1,9 @@
+package kitchenpos.table.validator;
+
+import java.util.function.Supplier;
+
+@FunctionalInterface
+public interface TableValidator {
+
+    Supplier<Boolean> validate(final Long orderTableId);
+}
