--- src/main/java/kitchenpos/ui/TableGroupRestController.java
@@ -1,33 +1,0 @@
-package kitchenpos.ui;
-
-import kitchenpos.application.TableGroupService;
-import kitchenpos.dto.TableGroupResponse;
-import kitchenpos.dto.TableGroupSaveRequest;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.net.URI;
-
-@RestController
-@RequestMapping("/api/table-groups")
-public class TableGroupRestController {
-
-    private final TableGroupService tableGroupService;
-
-    public TableGroupRestController(final TableGroupService tableGroupService) {
-        this.tableGroupService = tableGroupService;
-    }
-
-    @PostMapping
-    public ResponseEntity<TableGroupResponse> create(@RequestBody final TableGroupSaveRequest request) {
-        final TableGroupResponse created = tableGroupService.create(request);
-        final URI uri = URI.create("/api/table-groups/" + created.getId());
-        return ResponseEntity.created(uri).body(created);
-    }
-
-    @DeleteMapping("/{tableGroupId}")
-    public ResponseEntity<Void> ungroup(@PathVariable final Long tableGroupId) {
-        tableGroupService.ungroup(tableGroupId);
-        return ResponseEntity.noContent().build();
-    }
-}
