--- src/main/java/kitchenpos/ui/TableRestController.java
@@ -1,48 +1,0 @@
-package kitchenpos.ui;
-
-import kitchenpos.application.TableService;
-import kitchenpos.dto.OrderTableChangeEmptyRequest;
-import kitchenpos.dto.OrderTableChangeNumberOfGuestsRequest;
-import kitchenpos.dto.OrderTableResponse;
-import kitchenpos.dto.OrderTableSaveRequest;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.net.URI;
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/tables")
-public class TableRestController {
-
-    private final TableService tableService;
-
-    public TableRestController(final TableService tableService) {
-        this.tableService = tableService;
-    }
-
-    @PostMapping
-    public ResponseEntity<OrderTableResponse> create(@RequestBody final OrderTableSaveRequest request) {
-        final OrderTableResponse created = tableService.create(request);
-        final URI uri = URI.create("/api/tables/" + created.getId());
-        return ResponseEntity.created(uri).body(created);
-    }
-
-    @GetMapping
-    public ResponseEntity<List<OrderTableResponse>> list() {
-        return ResponseEntity.ok().body(tableService.list());
-    }
-
-    @PutMapping("/{orderTableId}/empty")
-    public ResponseEntity<OrderTableResponse> changeEmpty(@PathVariable final Long orderTableId,
-                                                          @RequestBody final OrderTableChangeEmptyRequest request) {
-        return ResponseEntity.ok().body(tableService.changeEmpty(orderTableId, request));
-    }
-
-    @PutMapping("/{orderTableId}/number-of-guests")
-    public ResponseEntity<OrderTableResponse> changeNumberOfGuests(
-            @PathVariable final Long orderTableId,
-            @RequestBody final OrderTableChangeNumberOfGuestsRequest request) {
-        return ResponseEntity.ok().body(tableService.changeNumberOfGuests(orderTableId, request));
-    }
-}
