+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -16,14 +16,14 @@ import java.util.List;
 import kitchenpos.menu.domain.MenuDao;
 import kitchenpos.menugroup.domain.MenuGroupDao;
 import kitchenpos.order.domain.OrderDao;
-import kitchenpos.dao.OrderTableDao;
+import kitchenpos.table.domain.OrderTableDao;
 import kitchenpos.product.domain.ProductDao;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 import kitchenpos.product.domain.Product;
 import kitchenpos.order.dto.OrderChangeOrderStatusRequest;
 import kitchenpos.order.dto.OrderLineItemSaveRequest;
